documents-ocr-fn/resources/function/requirements.txt
matplotlib>=3.2.2
numpy==1.23.1
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.12.1
torchvision>=0.8.1
tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0
coremltools==6.3.0
tensorflow==2.8
tensorflowjs>=3.9.0
thop
coretex
pytesseract
easyocr
transformers
ultralytics
